Eliminación de algunos warnings Macro TL_TODO para sacar en los mensajes del compilador los mensajes TODO

+++ examples/math/rotations.cpp
@@ -59,9 +59,9 @@ int main(int argc, char** argv)
 
   // Paso de matriz de rotación a ángulos de Euler
   Matrix3f m;
-  m = AngleAxisf(0.25*TL_PI, Vector3f::UnitX()) *
-      AngleAxisf(0.5*TL_PI, Vector3f::UnitY()) *
-      AngleAxisf(0.33*TL_PI, Vector3f::UnitZ());
+  m = AngleAxisf(static_cast<float>(0.25*TL_PI), Vector3f::UnitX()) *
+      AngleAxisf(static_cast<float>(0.5*TL_PI), Vector3f::UnitY()) *
+      AngleAxisf(static_cast<float>(0.33*TL_PI), Vector3f::UnitZ());
 
   std::cout << "original rotation:" << std::endl;
   std::cout << m << std::endl << std::endl;
